steps towards being able to do a slick user dao
Former-commit-id: bae3249cd532893da9dd06ccc204193239610515

+++ src/mitll/langtest/shared/flashcard/SetScore.java
@@ -13,5 +13,5 @@ public interface SetScore {
   float getAvgScore();
   float getCorrectPercent();
 
-  long getUserid();
+  int getUserid();
 }
